test(widget): migrate Tabs, Tooltip, Input tests to integration tests
Add tests for: Tabs, Tooltip, Input

- Tabs: new, builder, navigation, handle_key, render, selected_label
- Tooltip: visibility, delay, render
- Input: new, value, type_char, backspace, cursor_movement, select_all, undo/redo

Tests accessing private fields remain inline in source files.
Total widget integration tests: 127
